requirements.txt
Django==1.8.6
djangorestframework==3.3.1
psycopg2==2.6.1
requests==2.8.1
